Move glucose-specific functions to glucose module
