+++ aps/lib/infinity_aps/glucose/monitor.ex
@@ -17,9 +17,29 @@ defmodule InfinityAPS.Glucose.Monitor do
 
   def write_oref0(entries, local_timezone) do
     entries
-    |> Enum.filter(&InfinityAPS.filter_sgv/1)
-    |> Enum.map(fn entry -> InfinityAPS.map_sgv(entry, local_timezone) end)
+    |> Enum.filter(&filter_sgv/1)
+    |> Enum.map(fn entry -> map_sgv(entry, local_timezone) end)
     |> Poison.encode!()
     |> InfinityAPS.write_data("cgm.json")
+  end
+
+  @doc """
+  Returns true if the input is a sensor_glucose_value.
+  Otherwise, it returns false.
+
+  Used for sgv filtering with Enum.filter/1.
+  """
+  def filter_sgv({:sensor_glucose_value, _}), do: true
+  def filter_sgv(_), do: false
+
+  @doc """
+  Converts the sensor_glucose_value `entry_data` into an sgv entry, taking into account
+  the `local_timezone`.
+  """
+  def map_sgv({:sensor_glucose_value, entry_data}, local_timezone) do
+    date_with_zone = Timex.to_datetime(entry_data.timestamp, local_timezone)
+    date = DateTime.to_unix(date_with_zone, :milliseconds)
+    date_string = Timex.format!(date_with_zone, "{ISO:Extended:Z}")
+    %{type: "sgv", sgv: entry_data.sgv, date: date, dateString: date_string}
   end
 end
